Small fixes and feature additions to stringdb, render_flat
- Fixed entry count logging with stringdb on init
- Added color control sequence to render_flat
- (TODO : migrate formats to stringdb structures to prevent problems like the typewriter process with format sequences)

Right now, the env text output will appear weird, as halfway through the printing process only half of a format string will be sent to render_flat (as the string entry is iteratively stepped through).
This will cause the render_flat text rendering procedure to panic and abort the render process.
A convienient way to avoid this is to store format indicators outside of the string text and to pass them as a seperate structure (a linked list with target indices and formats?)
These should be parsed at load-time into the stringdb to keep things simple.

After this, we will be able to easily render formatted strings without complications like iteratively stepping through half-format sequences.