Thoroughly broke everything
Altered coordinate system to use all but the last nybble for block indices and the last one for a 4-bit index into each "pixel".
Added some new types and started migrating things over.

There are several things which need to be completed --
First, the rendering system needs to be migrated to use the new coordinate system. VBOs still operate in a float-range as they should,
as all rendering will still use floating-point vertices, colors, etc. What needs to change is the transformation process from world coordinates to
screenspace coordinates and this will have to take place in the camera. Unfortunately due to precision errors we will NOT want to do this with a single
orthographic projection so we will not be able to multiply a single matrix. We will have to have an I/O function to translate world points and will have to start using it in render.c.

Second, the world loading system is heavily flawed. The load process takes place in tds_engine while the segment reduction and hblock generation process takes place in tds_world. As a result, the data cannot be streamed from the file directly into these algorithms and the entire world is loaded into a large 8-bit buffer and then passed in its entirety through tds_world_load. This is a huge waste of memory and likely brutal on systems with less RAM. With the new coordinate system it is completely infeasible as worlds no longer have a variable size; they will always span the entire (2^32)^2 size which means storing the entire buffer would take several hundred thousand terabytes of RAM. We need to alter the loading process to allow streaming to the hblock generation process; perhaps the FD itself could be passed to the streamgen function while world data is being loaded. This might not look pretty but it would solve a LOT of problems.

Several other systems need to be migrated as well. hunter-game will definitely not compile with this branch.